+++ app/assets/javascripts/source/app.js
@@ -2,7 +2,7 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 
 import Header from './header.js';
-import Questionnaires from './form.js';
+import Questionnaires from './questionnaires.js';
 
 Header();
 Questionnaires();
