+++ app/assets/javascripts/source/questionnaires.js
@@ -1,0 +1,5 @@
+import React from 'react';
+import ReactDOM from 'react-dom';
+import Questionnaires from './questionnaires/questionnaires.js';
+
+export default () => { ReactDOM.render(<Questionnaires />, document.getElementById('form')); };
