+++ app/assets/javascripts/source/questionnaires/option.js
@@ -1,0 +1,16 @@
+import React from 'react';
+
+export default class Option extends React.Component {
+  constructor(props) {
+    super(props);
+
+  }
+
+  render() {
+    return (
+      <option value={this.props.value}>
+        {this.props.text}
+      </option>
+    )
+  }
+}
